Fix build errors
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@9556 3db68df2-c116-0410-a063-a993310a9797

+++ source/ps/GameSetup/HWDetect.cpp
@@ -182,7 +182,7 @@ void RunHardwareDetection()
 
 	scriptInterface.SetProperty(settings.get(), "gfx_card", gfx::CardName());
 	scriptInterface.SetProperty(settings.get(), "gfx_drv_ver", gfx::DriverInfo());
-	scriptInterface.SetProperty(settings.get(), "gfx_mem", gfx::MemorySizeMiB());
+	scriptInterface.SetProperty(settings.get(), "gfx_mem", (u32)gfx::MemorySizeMiB());
 
 	scriptInterface.SetProperty(settings.get(), "snd_card", std::wstring(snd_card));
 	scriptInterface.SetProperty(settings.get(), "snd_drv_ver", std::wstring(snd_drv_ver));
